backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0
django-cors-headers>=4.3,<4.4
django-filter>=24.2,<25.0
python-dotenv>=1.0,<1.1
drf-spectacular>=0.27,<0.28
